move all Navigation code into newly created folder.
