+++ src/components/App.js
@@ -23,14 +23,8 @@ import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
 import Button from 'react-bootstrap/Button';
 import { LinkContainer } from 'react-router-bootstrap';
 import './App.css';
-import { NavLink, Switch, Route } from 'react-router-dom'
+import { Navigation, Main } from './Navigation'
 
-
-// const Home = () => <span>Home</span>;
-
-// const About = () => <span>About</span>
-
-// const Users = () => <span>Users</span>
 
 class App extends React.Component {
 
@@ -176,45 +170,7 @@ class App extends React.Component {
   }
 }
 
-const Navigation = () => (
-  <nav>
-    <ul>
-      <li><NavLink exact activeClassName="current" to='/'>Home</NavLink></li>
-      <li><NavLink exact activeClassName="current" to='/about'>About</NavLink></li>
-      <li><NavLink exact activeClassName="current" to='/contact'>Contact</NavLink></li>
-    </ul>
-  </nav>
-);
 
-const Home = () => (
-  <div className='home'>
-    <h1>Welcome to my portfolio website</h1>
-    <p> Feel free to browse around and learn more about me.</p>
-  </div>
-);
-
-const About = () => (
-  <div className='about'>
-    <h1>About Me</h1>
-    <p>Ipsum dolor dolorem consectetur est velit fugiat. Dolorem provident corporis fuga saepe distinctio ipsam? Et quos harum excepturi dolorum molestias?</p>
-    <p>Ipsum dolor dolorem consectetur est velit fugiat. Dolorem provident corporis fuga saepe distinctio ipsam? Et quos harum excepturi dolorum molestias?</p>
-  </div>
-);
-
-const Contact = () => (
-  <div className='contact'>
-    <h1>Contact Me</h1>
-    <p>You can reach me via email: <strong>hello@example.com</strong></p>
-  </div>
-);
-
-const Main = () => (
-  <Switch>
-    <Route exact path='/' component={Home}></Route>
-    <Route exact path='/about' component={About}></Route>
-    <Route exact path='/contact' component={Contact}></Route>
-  </Switch>
-);
 
 App.propTypes = {
   display: PropTypes.object,
